Extract Piskel and Framesheet models from models __init__
